+++ tests/testthat/test-vetted_models-glm.R
@@ -26,7 +26,6 @@ skip_on_cran()
 skip_if(!isTRUE(as.logical(Sys.getenv("CI"))))
 skip_if_not(broom.helpers::.assert_package("car", pkg_search = "gtsummary", boolean = TRUE))
 skip_if_not(broom.helpers::.assert_package("survival", pkg_search = "gtsummary", boolean = TRUE))
-library(dplyr)
 
 # glm() ------------------------------------------------------------------------
 test_that("vetted_models glm()", {