+++ R/as_flex_table.R
@@ -350,7 +350,7 @@ table_styling_to_flextable_calls <- function(x, ...) {
   if (!is.null(x$table_styling$source_note)) {
     flextable_calls[["source_note"]] <-
       expr(
-        flextable::footnote(value = flextable::as_paragraph(!!x$table_styling$source_note), ref_symbols = " ")
+        flextable::add_footer_lines(value = flextable::as_paragraph(!!x$table_styling$source_note))
       )
   }